# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Service:
    properties:
      api_type:
        description: HTTP API, gRPC, MQTT, TCP, DNS, Ping
        type: string
      check_interval:
        description: in seconds
        type: integer
      created_at:
        type: string
      description:
        type: string
      dns_domain_name:
        description: Specific fields for DNS
        type: string
      group:
        type: string
      grpc_auth:
        type: string
      grpc_host:
        description: Specific fields for gRPC
        type: string
      grpc_port:
        type: integer
      grpc_proto:
        type: string
      http_expected_status:
        type: integer
      http_health_url:
        type: string
      http_method:
        description: Specific fields for HTTP API
        type: string
      id:
        type: string
      label:
        type: string
      mqtt_auth:
        type: string
      mqtt_host:
        description: Specific fields for MQTT
        type: string
      mqtt_port:
        type: integer
      mqtt_qos:
        type: integer
      mqtt_topic:
        type: string
      name:
        type: string
      ping_host:
        description: Specific fields for Ping
        type: string
      tcp_host:
        description: Specific fields for TCP
        type: string
      tcp_port:
        type: integer
      timeout:
        description: in seconds
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Monitron Backend.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Monitron Backend API
  version: "1.0"
paths:
...
  /services:
    get:
      description: Retrieve a list of all monitoring services
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Service'
            type: array
        "500":
          description: 'error": "Could not retrieve services'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all services
      tags:
      - Services
    post:
      consumes:
      - application/json
      description: Create a new monitoring service
      parameters:
      - description: Service object to be created
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/models.Service'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Service'
        "400":
          description: 'error": "Cannot parse JSON'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not create service'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new service
      tags:
      - Services
  /services/{id}:
    delete:
      description: Delete a monitoring service by its ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: 'error": "Invalid service ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error": "Service not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not delete service'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a service
      tags:
      - Services
    get:
      description: Retrieve a single monitoring service by its ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Service'
        "400":
          description: 'error": "Invalid service ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error": "Service not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not retrieve service'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get service by ID
      tags:
      - Services
    put:
      consumes:
      - application/json
      description: Update details of an existing monitoring service by its ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Service object with updated fields
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/models.Service'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Service'
        "400":
          description: 'error": "Invalid service ID" or "Cannot parse JSON'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error": "Service not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not update service'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update an existing service
      tags:
      - Services
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"